refactor: Update ThemeTemplate functions to improve code organization and readability
